+++ datacraft/server.py
@@ -1,6 +1,7 @@
 """
 Light weight module for running a Flask Server that returns data from a generator.
 """
+import time
 from typing import Generator
 import logging
 import flask
@@ -18,12 +19,16 @@ class _Server:
     def __init__(self,
                  generator: Generator,
                  endpoint: str,
+                 port: int,
                  data_is_json: bool,
-                 count_supplier: datacraft.ValueSupplierInterface):
+                 count_supplier: datacraft.ValueSupplierInterface,
+                 delay: float = None):
         self.generator = generator
         self.endpoint = endpoint
+        self.port = port
         self.data_is_json = data_is_json
         self.count_supplier = count_supplier
+        self.delay = delay
         self.call_number = 0
 
     def callback(self):
@@ -40,6 +45,8 @@ class _Server:
         except StopIteration:
             _log.warning('No more iterations available')
             return flask.Response(None, status=204)
+        if self.delay:
+            time.sleep(self.delay)
         if self.data_is_json:
             return flask.jsonify(data)
         # this may already be json or templated data
@@ -52,14 +59,18 @@ class _Server:
         """ run the Flask app """
         app = flask.Flask(__name__)
         _log.info('Adding endpoint to server: %s', self.endpoint)
-        app.add_url_rule(self.endpoint, view_func=self.callback)
-        app.run()
+        app.add_url_rule(self.endpoint,
+                         view_func=self.callback,
+                         methods=['POST', 'GET'])
+        app.run(port=self.port)
 
 
 def run(generator: Generator,
         endpoint: str,
+        port: int,
         data_is_json: bool,
-        count_supplier: datacraft.ValueSupplierInterface):
+        count_supplier: datacraft.ValueSupplierInterface,
+        delay: float = None):
     """
     Runs a light weight Flask server with data returned by the provided generator served at each subsequent call to
     the provided endpoint. End point should start with /. When StopIteration encountered, returns a 204 status code.
@@ -67,8 +78,10 @@ def run(generator: Generator,
     Args:
         generator: that provides response data as dictionary
         endpoint: to serve data on i.e. /data
+        port: to serve data on e.g. 8080
         data_is_json: if the data should be returned as JSON, default is as string
         count_supplier: supplies the number of values to generate for each call
+        delay: number of seconds to pause between request and response
     """
-    server = _Server(generator, endpoint, data_is_json, count_supplier)
+    server = _Server(generator, endpoint, port, data_is_json, count_supplier, delay)
     server.run()